ref(test): Convert eventsV2/results test to rtl and tsc
